Fix MLX seed stepper crash: clamp UInt64 to UInt32 range
The SwiftUI Stepper with 0...UInt64.max causes 'Distance is not representable in Int'
when computing stride distances. Changed to 0...UInt32.max which is still 4B values
but avoids the Int overflow. Also added clamping to all MLXSettings numeric
properties via clamped(to:) extension to prevent out-of-bounds UserDefaults values.

+++ App/SwairmApp/ContentView.swift
@@ -108,7 +108,7 @@ struct ContentView: View {
                                 value: $mlxController.settings.loraAlpha, in: 1...64, step: 1)
                             .disabled(isRunning)
                         Stepper("Seed: \(mlxController.settings.seed)",
-                                value: $mlxController.settings.seed, in: 0...UInt64.max, step: 1)
+                                value: $mlxController.settings.seed, in: 0...UInt32.max, step: 1)
                             .disabled(isRunning)
                     }
